feat: move /recommendations to /prompts
The book prompts page displays prompts not recommendations, so change the page routing accordingly.
