+++ src/app/(protected)/prompts/BookPromptContext.ts
