+++ src/app/(protected)/prompts/page.tsx
@@ -4,7 +4,6 @@ import BookPromptsTable from '@/components/book-prompt/BookPromptsTable';
 import { Button } from '@/components/ui/button';
 import { getBookPrompts } from '@/lib/api';
 import { BookPromptTable } from '@/types/BookPromptTable';
-import { BookCopyIcon } from 'lucide-react';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { useCallback, useEffect, useState } from 'react';
@@ -24,14 +23,11 @@ export default function RecommendationsPage() {
 
   return (
     <div>
-      <h1 className="flex gap-2 items-center mb-10">
-        <BookCopyIcon size={18} />
-        Recommendations
-      </h1>
       <div className="grid gap-4">
+        <h1>Prompts and Recommendations</h1>
         <div className="flex justify-end">
-          <Link href="/recommendations/new">
-            <Button>New</Button>
+          <Link href="/prompts/new">
+            <Button>New Prompt</Button>
           </Link>
         </div>
         <BookPromptsTable bookPrompts={bookPrompts} linkPathname={pathname} />
